requirements.txt
ConfigParser==6.0.0
dagshub==0.3.9
numpy==1.26.3
pandas==2.2.0
Pillow==10.1.0
protobuf==3.12.4
psycopg2_binary==2.9.9
pycocotools==2.0.7
python-dotenv==1.0.1
toml==0.10.2
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
zenml==0.56.1
pytest-xdist
pytest-asyncio
pytest-cov
mlflow==2.10.0